requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn==0.34.0

# Database
aiosqlite==0.20.0
lancedb==0.17.0
pyarrow>=14.0.0

# Document Parsing
pymupdf==1.25.1
python-docx==1.1.2
python-pptx==1.0.2
beautifulsoup4==4.12.3
lxml==5.3.0

# HTTP Client (for NIM API)
httpx==0.28.1

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18

# Tokenizer for accurate token counting
tokenizers>=0.15.0
